refactor mcp swift dto parser
